Correctly catching console commands
Currently correctly catching console probing commands, particularly
initialization (0x00 then a 1) and read ((0x400300 then a 1) commands.

* Hardcoded the assembly for data wait low and data line wait high for
every single phase of the checking, as far as I could research can't
store a jump label inside a register then jump to that register's value
* Currently NOT responding to them at all. Simply can catch them
regularly. Also not sure how to respond to the initialization command
(no documentation found on it nor did my oscilliscope or Arduino reading
methods turn up anything in that single instant that's extremely hard to
capture without the right tool)

Issues:
* Hardcoded for my specific initialization (0x00 then a single high bit)
and read (0x400300 then a single high bit) probe commands. May be
different on the Wii U, Wii, Gamecube, or within even consoles. OR it
could possibly be a parity/check bit (as there are 8 0's in a row in a
single byte), which is sorta odd as there are 12 0's in a row in the
read command. Doubts due to the old GC documentation that everyone
references, where the last one is actually a little earlier (and all of
the commands are nice and complete bytes)